requirements.txt
aiocoap==0.4.7
attrs==23.1.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
